[test] 자동차 테스트 다시 작성
