+++ src/main/java/study/step3/Car.java
@@ -4,15 +4,19 @@ import study.step2.StringValidator;
 
 public class Car{
     private static final int START_POSITION = 1;
+
     private String carName;
+
     private int position;
+
     public Car(String carName){
         StringValidator.checkCarNameSyntax(carName);
         this.carName = carName;
         this.position = START_POSITION;
     }
+
     public void move(MoveStrategy moveStrategy){
-        if(moveStrategy.isMove(carName.length())){
+        if(moveStrategy.isMove()){
             position++;
         }
     }
@@ -20,6 +24,7 @@ public class Car{
     public int getPosition(){
         return position;
     }
+
     public String getCarName(){
         return carName;
     }
