--- src/main/java/study/step3/CarMoveByNameLengthStrategy.java
@@ -1,10 +1,0 @@
-package study.step3;
-
-import study.step2.StringValidator;
-
-public class CarMoveByNameLengthStrategy implements MoveStrategy{
-    @Override
-    public boolean isMove(int condition) {
-        return condition == StringValidator.CAR_NAME_LENGTH_MOVE_CONDITION;
-    }
-}
