+++ src/main/java/study/step3/MoveStrategy.java
@@ -4,5 +4,5 @@ package study.step3;
 * 여러가지 다른 전략들이 생길 수 있음!!
 * */
 public interface MoveStrategy {
-    boolean isMove(int condition);
+    boolean isMove();
 }
