+++ src/main/java/study/step3/RandomMoveStrategy.java
@@ -4,7 +4,7 @@ import java.util.Random;
 
 public class RandomMoveStrategy implements MoveStrategy{
     @Override
-    public boolean isMove(int condition) {
-        return new Random().nextInt(NumberChecker.RANDOM_BOUND) >= condition;
+    public boolean isMove() {
+        return new Random().nextInt(NumberChecker.RANDOM_BOUND) >= NumberChecker.CONDITION_OF_START;
     }
 }
